feat: add helper for getting activity timeouts
