+++ olake-workers/k8s/activities/activities.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"time"
 
 	"go.temporal.io/sdk/activity"
 
@@ -14,6 +13,7 @@ import (
 	"olake-ui/olake-workers/k8s/pods"
 	"olake-ui/olake-workers/k8s/shared"
 	"olake-ui/olake-workers/k8s/utils/filesystem"
+	"olake-ui/olake-workers/k8s/utils/helpers"
 )
 
 // Activities holds the dependencies for activity functions
@@ -32,28 +32,6 @@ func NewActivities(jobService service.JobDataService, podManager *pods.K8sPodMan
 	}
 }
 
-// getTimeoutFromContext extracts timeout from Temporal activity context deadline
-// Pod timeout is set to activity timeout minus 2 minutes to allow cleanup time
-func (a *Activities) getTimeoutFromContext(ctx context.Context) time.Duration {
-	deadline, ok := ctx.Deadline()
-	if !ok {
-		// Fallback to 30 minutes if no deadline set
-		return time.Minute * 30
-	}
-
-	activityTimeout := time.Until(deadline)
-
-	// Pod timeout = activity timeout minus 2 minutes cleanup buffer
-	// This ensures pods terminate before Temporal activity timeout
-	podTimeout := activityTimeout - (2 * time.Minute)
-
-	// Minimum pod timeout of 1 minute
-	if podTimeout < time.Minute {
-		return time.Minute
-	}
-
-	return podTimeout
-}
 
 // DiscoverCatalogActivity discovers data source catalog using Kubernetes Pod
 func (a *Activities) DiscoverCatalogActivity(ctx context.Context, params shared.ActivityParams) (map[string]interface{}, error) {
@@ -76,7 +54,7 @@ func (a *Activities) DiscoverCatalogActivity(ctx context.Context, params shared.
 		Configs: []shared.JobConfig{
 			{Name: "config.json", Data: params.Config},
 		},
-		Timeout: a.getTimeoutFromContext(ctx),
+		Timeout: helpers.GetActivityTimeout("discover"),
 	}
 
 	// Execute discover operation by creating K8s pod, wait for completion, retrieve results from streams.json file
@@ -114,7 +92,7 @@ func (a *Activities) TestConnectionActivity(ctx context.Context, params shared.A
 		Configs: []shared.JobConfig{
 			{Name: "config.json", Data: params.Config},
 		},
-		Timeout: a.getTimeoutFromContext(ctx),
+		Timeout: helpers.GetActivityTimeout("test"),
 	}
 
 	// Execute check operation by creating K8s pod, wait for completion, retrieve results from pod logs
@@ -170,7 +148,7 @@ func (a *Activities) SyncActivity(ctx context.Context, params shared.SyncParams)
 			{Name: "writer.json", Data: jobData.DestConfig},
 			{Name: "state.json", Data: stateData},
 		},
-		Timeout: a.getTimeoutFromContext(ctx),
+		Timeout: helpers.GetActivityTimeout("sync"),
 	}
 
 	// Execute sync operation by creating K8s pod, wait for completion, retrieve results from state.json file
